+++ src/main/java/eu/lucaventuri/fibry/MiniQueue.java
@@ -4,32 +4,50 @@ import java.util.concurrent.LinkedBlockingDeque;
 import java.util.concurrent.TimeUnit;
 
 public interface MiniQueue<T> {
-    boolean add (T message);
-    default boolean offer (T message) {
+    boolean add(T message);
+
+    public boolean offer(T message, long timeout, TimeUnit unit) throws InterruptedException;
+
+    default boolean offer(T message) {
         try {
             return add(message);
-        } catch(Exception e) {
+        } catch (Exception e) {
             return false;
         }
     }
-    T take () throws InterruptedException ;
+
+    T take() throws InterruptedException;
+
     T poll(long timeout, TimeUnit unit) throws InterruptedException;
+
     void clear();
+
     int size();
+
     T peek();
 
-    /** LinkedBlockingDeque implementing MiniQueue */
+    /**
+     * LinkedBlockingDeque implementing MiniQueue
+     */
     static <T> MiniQueue<T> blocking() {
-        class MiniQueuedBlockingQueue<T> extends LinkedBlockingDeque<T> implements MiniQueue<T> {};
+        class MiniQueuedBlockingQueue<T> extends LinkedBlockingDeque<T> implements MiniQueue<T> {
+        }
 
         return new MiniQueuedBlockingQueue<>();
     }
 
-    /** MiniQueue dropping all the messages */
+    /**
+     * MiniQueue dropping all the messages
+     */
     static <T> MiniQueue<T> dropping() {
         return new MiniQueue<T>() {
             @Override
             public boolean add(T message) {
+                return false;
+            }
+
+            @Override
+            public boolean offer(T message, long timeout, TimeUnit unit) throws InterruptedException {
                 return false;
             }