+++ RELEASE_NOTES.md
@@ -1,3 +1,6 @@
-- Changed Autocloseable behaviour of group leader: now it sends a poison pill and waits
-- Added method sendPoisonPillOnExit()
+- Improved support for actors with bounded queue:
+  - execute() now waits indefinitely, instead of throwing an exception
+  - sendPoisonPill() now waits indefinitely, instead of simply returning false
+  - added execAsyncTimeout(), that does not throw an exception
+- Version 1.0.9